chore(wizard-refactor): extract resolvers/infer-type-from-name
Pure regex helper, no deps. Dispatcher tests 11/11 + lint test 17/17 green.
